consolidate house styling
