+++ frontend/src/components/House/HouseForm.js
@@ -1,7 +1,6 @@
 import React, {Component} from 'react';
 import PropTypes from 'prop-types';
-import {css} from 'aphrodite';
-import styles from './styles';
+import {StyleSheet, css} from 'aphrodite';
 
 export default class HouseForm extends Component {
   static propTypes = {
@@ -31,12 +30,12 @@ export default class HouseForm extends Component {
 
         <label className={css(styles.left_grid)}>Street</label>
         <input name="location.street" type="text" placeholder="Street"
-          className={css(styles.right_gred)}
+          className={css(styles.right_grid)}
           value={house.location.street} onChange={onFieldChange} />
 
-        <label className={css(styles.left_gred)}>postcode</label>
+        <label className={css(styles.left_grid)}>Post Code</label>
         <input name="location.postCode" type="text" placeholder="Post Code"
-          className={css(styles.right_gred)}
+          className={css(styles.right_grid)}
           value={house.location.postCode} onChange={onFieldChange} />
 
         <label className={css(styles.left_grid)}>Description</label>
@@ -53,3 +52,22 @@ export default class HouseForm extends Component {
     );
   };
 }
+
+const styles = StyleSheet.create({
+  form: {
+    margin: "10px auto",
+    maxWidth: "600px",
+    display: "grid",
+    alignItems: "baseline",
+    gridTemplateColumns: "100px 1fr",
+    gridGap: "8px"
+  },
+  left_grid: {
+    lineHeight: "1.5em",
+    textAlign: "right",
+    gridColumn: "1 / 2"
+  },
+  right_grid: {
+    gridColumn: "2/3"
+  }
+});
